+++ src/libpurple/protocols/msn/nexus.h
@@ -206,7 +206,7 @@ struct _MsnNexus
 	char *cipher;
 	char *secret;
 	MsnTicketToken *tokens;
-	gsize token_len;
+	int token_len;
 };
 
 void msn_nexus_connect(MsnNexus *nexus);
